Implemented a basic neural net with theano (no learning)
It doesn't learn yet, so it doesn't even really qualify as a neural net.
Right now it just builds a layered network of perceptrons with specified
numbers and sizes of layers and random weights. I'll make it actually
learn things soon; this was mostly just familiarizing myself with the
library.